+++ Changelog.md
@@ -1,9 +1,10 @@
 # Changelog for nextcloud api
 
 ## Version 14.1.0
-- 2025-10-21
+- 2025-10-23
   - Add HTTP proxy support (via system variables `https.proxyHost` and `https.proxyPort`)
-    
+  - Add User properties: lastLogin, backend, language, locale, and subAdminGroups
+
 ## Version 14.0.0
 - 2025-10-21
   - Bump required java version from 8 to 11+ (Thanks to kindlich)